--- docs/src/pages/components/dropdown.js
@@ -1,1 +1,0 @@
-export default from '../../content/components/dropdown';
