+++ docs/src/pages/components/select.js
@@ -1,0 +1,1 @@
+export default from '../../content/components/select';
