+++ src/Components/Select/Select.test.js
@@ -1,10 +1,10 @@
 import React from 'react';
-import { Dropdown } from '../..';
+import { Select } from '../..';
 
 import ShallowRenderer from 'react-test-renderer/shallow';
 const renderer = new ShallowRenderer();
 
 test('Dropdown Renders', () => {
-  const tree = renderer.render(<Dropdown />);
+  const tree = renderer.render(<Select />);
   expect(tree).toMatchSnapshot();
 });
